ENGR00156967 Gallery can't delete buckets.
Root cause: When delete buckets, the MediaProvider couldn't get the
correct database for the delete parameters didn't contain any information
about whick database it was in.
The way to fix it: Do the delete command in all the databases. For the
bucket id was generated through the path, the delete command would delete
the right files in right database.

Signed-off-by: YangYonggang <b31664@freescale.com>