ENGR00216809 Add null pointer check before do addStorage()
Before /mnt/sdcard primary storage mounted, the MtpStorage
instance is not created, and this cause AndroidRuntime
crash when calling the storage's method with PTP enabled on boot.
So we must add a sanity check here, do not add storage for
null one, until the storage is mounted, added by volume listener.

Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>